refactor: move duplicate constants between `scripts` and `src` to `src` (#237)
